+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## master
+
+  * Add commands for Sequence, Scene, Area, Curtain Switch, Floor Heating,
+    Panel Control, Sensors and Read Temperature sections of
+    system specification
+  * Add missing Channels and Universal Switch commands
+
 ## 0.3.0
 
   * URL syntax for configuration